Code refactor [no new functionality]
