+++ resources/grocery-list.txt
@@ -1,3 +1,2 @@
-        first thing
-        second thing
-        
+stuff here
+how about thing two
